convert madis to python3

+++ iis-3rdparty-madis/src/main/resources/eu/dnetlib/iis/3rdparty/scripts/madis/lib/sqlitetypes.py
@@ -1,6 +1,6 @@
 def getElementSqliteType(element):
     elemtype = type(element)
-    if elemtype is int or elemtype is long:
+    if elemtype is int or elemtype is int:
         return "integer"
     if elemtype is float:
         return "real"
